Use Disconnected message to acknowledge quitting
The message is used to tell the client that he has been disconnected
from the server. It can be improved later to include reason for
disconnect so server can send this message to all the clients before
shutting down and clients will be able to see what was the reason.