Split Account and Player concepts
Account has login and password, player has name and other game-specific
attributes. One account can have many players. When a client signs in,
they supply login, password, and player name. If account with that login
does not exist - it's created. If player with this name does not exist -
it's created. If account exists but password is incorrect, client is
disconnected with respective reason.