Add Say message for basic notification support
Say message can be used to send notifications from server to clients.

Optionally it can be used by clients to send chat messages to server
that will be then broadcasted to all the clients nearby.

Maybe it's a good idea to implement chat system as a separate
microservice subsystem though. But still Say message can be used for
some scenarios when something needs to be said in the common message
processing queue of the game.