Update state every time update comes in
We can't postpone updating when update status comes in. Whenever update
status comes in - we need to update the current state because we don't
know what was the reason of that update. Maybe it was some neighbor
coming in.

We can of course compare previous neighbor list with current neighbor
list and build sophisticated change detection engine later.

But for now it's easier to update everything every time.

This introduces an issue though: every time we send Move command (every
10 steps) - this results in Update of current status sent back to
client. And when we rebuild the current state - we can set a different
text for current process of moving on the road, which is wrong and quite
annoying.

We need to not change text, or not even change the typer, when Distance
did not change. Maybe even cache forward/backward typers so we have the
same text when turning around.