Send road direction information with road from server
Client doesn't know which road belongs to which location. He needs to
know which direction of the road he needs to show to the user.

Maybe better approach would be to make all roads one-way. And when
turning back - to move to another road (with another ID). This will
eliminate all issues and weird code, but will introduce another issue of
inconsistent invariants of the road (road should swap between only two
specific locations).

This however can be made into a feature: what if there's a road, on
which when you turn around many times - you get lost.. :)